Build documentation using Zola
Builds a documentation site using the static site generator Zola and the
Adidoks theme.

Makefile targets were added for convenience.

This is just a proof of concept to give an idea of what the
documentation could look like, as well as the ability to write blog
posts. The site has the minimum configuration required, and metadata
will certainly need to be adjusted before merging this PR.

Extra content examples can be found in `docs/themes/adidoks/content`.