Fix usage of repository tag in examples (#114)
Fix the usage of the `repository` tag in the example pipelines.

The scripts should be using `REPOSITORY` instead of `REGISTRY` when setting the `repository` tag in the cue evaluation. The scripts are setup to automatically add a random 16-character string to the end of the `REGISTRY` value to produce a randomized `REPOSITORY`.