Update validation (#154)
* Update validation

Update documentation and CI build to reflect a consistent mechanism for obtaining the image URL from the last pipeline run.

This looks at the last pipeline run and pulls out the entry from `taskResults` that has a `name` ending with `IMAGE_URL`.

The task run that produced the image can also be identified in a similar fashion, since not all pipeline runs produce the image in the last task run.

With the task run that produced the image identified, we can then use its name for validating the signature and attestation using Tekton CLI (v0.23.0+)

* Consistent env variable usage

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>