Check pipeline run status in CI builds
Pipeline runs can currently fail without causing the step to immediately fail. This allows following parts of the step to be executed, which may or may not actually fail the build.

This PR adds a check on the status of the last pipeline run in each step of the CI build.

Signed-off-by: Brad Beck <bradley.beck@gmail.com>